requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
python-dotenv
pydantic
duckduckgo-search
langgraph-checkpoint-sqlite